requirements.txt
# 核心科学计算库
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# 地理空间处理
shapely>=1.8.0
rasterio>=1.2.0
geopandas>=0.10.0

# 图形算法与路径规划
networkx>=2.6.0

# 强化学习框架
pettingzoo>=1.18.0
gymnasium>=0.26.0
ray[rllib]>=2.0.0

# 优化求解器
pulp>=2.6.0

# 可视化
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# 配置管理
pyyaml>=6.0
hydra-core>=1.1.0

# 数据处理与存储
pyarrow>=8.0.0
h5py>=3.6.0
lxml>=4.6.0

# 测试框架
pytest>=7.0.0
pytest-cov>=3.0.0

# 工具库
tqdm>=4.62.0
rich>=12.0.0

# 图像处理
pillow>=9.0.0
opencv-python>=4.5.0

# 哈希与压缩
zstandard>=0.17.0